Log version on startup
Since this doesn't run through any type of build system that could
automatically generate the version string I came up with this.

When taken directly from a branch in this repo it will be in the format
COMMIT-BRANCH where COMMIT is the short commit hash of the upstream
commit that was most recently merged, and BRANCH is the local branch.
The commit hash will be updated every time upstream is merged.

When taken from a release it will be the tag name. Release tags will be
pointed at a detached commit that changes the version string to the tag name.